backend/requirements.txt
Flask == 1.1.2
# Werkzeug == 1.0.1
# Flask-cors==3.0.10
joblib == 1.0.0
python-dotenv
nltk ==3.5
unidecode==1.2.0
waitress == 2.0.0
numpy ==1.19.4
pymongo
pandas == 1.1.4
dnspython
cython
Pillow
requests
regex
Flask-RESTful
pyyaml
fastapi
uvicorn
python-multipart
